Move band test classes to 'commands.band'
